requirements.txt
altair==5.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dominate==2.9.0
filelock==3.13.1
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
flatbuffers==23.5.26
fsspec==2023.12.2
gitdb==4.0.11
GitPython==3.1.40
gTTS==2.4.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
imageio==2.33.1
importlib-metadata==6.11.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
lazy_loader==0.3
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.16.3
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
protobuf==4.25.1
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.17.2
PyMatting==1.1.12
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
regex==2023.10.3
rembg==2.0.53
requests==2.31.0
rich==13.7.0
rpds-py==0.13.2
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
six==1.16.0
smmap==5.0.1
soupsieve==2.5
streamlit==1.29.0
sympy==1.12
tenacity==8.2.3
tifffile==2023.12.9
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.1.1
tornado==6.4
tqdm==4.66.1
transformers==4.36.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
visitor==0.1.3
watchdog==3.0.0
Werkzeug==3.0.1
zipp==3.17.0
